Refactored footer section
