+++ src/components/Skills/Skills.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import './Skills.css'
+import "@fortawesome/fontawesome-free/css/all.min.css"; // Import Font Awesome CSS
 
 export default function Skills() {
     const skills = [
